Do not touch source meta
Make a distinction between the source meta and the message meta, i.e.
parentCorrelationId and notifyProcessed are not properties of the source
but of the message (in the case of spawning). Transfer the (internal)
message properties using headers (`x-notify-processed` and
`x-parent-correlation-id`).

Also use the message property (AMQP native) correlationId as context
correlationId when available.